+++ lib/factorix/cache/s3.rb
@@ -26,7 +26,7 @@ module Factorix
     #   lock_timeout = 30
     class S3 < Base
       # @!parse
-      #   # @return [Dry::Logger::Dispatcher]
+      #   # @return [Factorix::Logger]
       #   attr_reader :logger
       include Import[:logger]