+++ lib/factorix/cli/commands/cache/stat.rb
@@ -20,7 +20,7 @@ module Factorix
         #     ...
         class Stat < Base
           # @!parse
-          #   # @return [Dry::Logger::Dispatcher]
+          #   # @return [Factorix::Logger]
           #   attr_reader :logger
           include Import[:logger]
           include Formatting