+++ lib/factorix/cli/commands/mod/enable.rb
@@ -11,7 +11,7 @@ module Factorix
           backup_support!
 
           # @!parse
-          #   # @return [Dry::Logger::Dispatcher]
+          #   # @return [Factorix::Logger]
           #   attr_reader :logger
           #   # @return [Factorix::Runtime]
           #   attr_reader :runtime